LPS-182283 Inline ConfigurationFieldOptionsProviderUtil into ConfigurationModelToDDMFormConverter
